bullets & shooting
Bullets are set to collide with Walls, Enemies and currently are set to free when hit. Bullets are designed the way that the damage they cause can be changed. BulletManager is parent to all bullets. When player shoots, signal in Player node is emitted to BulletManager with parameters position and direction.