peaks & templates
Peaks has been created and added to main scene, also peaks as template scene as has been create to add more such scenes with different setups and collision shapes, GoblinSpawner has also been separated to template and spawner that is in main. Spawner on main has different attached different script that prescribes 10 goblins to be spawned. In this manner, spawners and peaks will be created over the map and added to main scene.