panel labels
All item variables belongs to player, as well as functions that process using the items and creating some effects on player. Health function is done, speed and shots are prepared for later. In player, signal were created and these signals are emitted to Panel when certain amount is changed, panel then changes label of this amount item.